+++ cmd/infra_down.go
@@ -13,8 +13,11 @@ import (
 
 var infraDownCmd = &cobra.Command{
 	Use:   "down",
-	Short: "Derruba uma infraestrutura de desenvolvimento provisionada anteriormente",
-	Long:  "Derruba uma infraestrutura de desenvolvimento provisionada anteriormente",
+	Short: "Take down the provisioned development infrastructure",
+	Long:  "Take down the provisioned development infrastructure",
+	Example: `To destroy the locally provisioned cluster, run:
+
+    $ safira infra down`,
 	RunE:  runInfraDown,
 	SuggestionsMinimumDistance: 1,
 }