+++ cmd/kong_new.go
@@ -12,8 +12,11 @@ import (
 
 var kongNewCmd = &cobra.Command{
 	Use:     "new [TYPE] [NAME]",
-	Short:   "Cria a estrutura de um yaml para o CRD do Kong",
-	Long:    "Cria a estrutura de um yaml para o CRD do Kong",
+	Short:   "Create a yaml structure for the Kong CRD",
+	Long:    "Create a yaml structure for the Kong CRD",
+	Example: `To create a yaml file for kong, run:
+
+    $ safira kong new plugin plugin-name`,
 	PreRunE: preRunKongpluginNew,
 	RunE:    runKongpluginNew,
 }