+++ cmd/template_list.go
@@ -28,8 +28,11 @@ type TemplateInfo struct {
 var templateListCmd = &cobra.Command{
 	Use:     "list",
 	Aliases: []string{"ls"},
-	Short:   "Lista os templates oficiais do Vertigo iPaaS",
-	Long:    "Lista os templates oficiais do Vertigo iPaaS",
+	Short:   "Lists the official Vertigo iPaaS templates",
+	Long:    "Lists the official Vertigo iPaaS templates",
+	Example: `To list all the official templates, run:
+
+    $ safira template list`,
 	RunE:    runTemplateList,
 	SuggestionsMinimumDistance: 1,
 }