+++ cmd/template_pull.go
@@ -9,8 +9,11 @@ import (
 
 var templatePullCmd = &cobra.Command{
 	Use:   "pull",
-	Short: "Baixa os templates oficiais do Vertigo iPaaS",
-	Long:  "Baixa os templates oficiais do Vertigo iPaaS",
+	Short: "Pull out the official Vertigo iPaaS templates",
+	Long:  "Pull out the official Vertigo iPaaS templates",
+	Example: `To pull the official templates, run:
+
+    $ safira template pull`,
 	RunE:  runTemplatePull,
 }